refactor: 관리자 페이지 리팩토링
