+++ src/components/Loading/Loading.module.css
@@ -6,7 +6,7 @@
   justify-content: center;
   align-items: center;
   z-index: 99;
-  background-color: rgba(0, 0, 0, 0.5);
+  background-color: rgba(0, 0, 0, 0.2);
 }
 
 .spinner {
