+++ src/pages/Admin/Banner/BannerManagement.jsx
@@ -1,15 +1,19 @@
-import { Button, Modal, ModalContent, ModalTrigger } from "@/components";
+import {
+  Button,
+  Loading,
+  Modal,
+  ModalContent,
+  ModalTrigger,
+} from "@/components";
 import { useBanner } from "@/hooks";
-import { formatDate } from "date-fns";
-import styles from "./BannerManagement.module.css";
-import ModifyBanner from "./ModifyBanner";
 import AddBanner from "./AddBanner";
-import DeleteModal from "@/components/Modal/ModalContents/DeleteModal";
+import styles from "./BannerManagement.module.css";
+import BannerTable from "./BannerTable";
 
 const BannerManagement = () => {
   const { allBanner, addBanner, updateBanner, deleteBanner } = useBanner();
 
-  if (!allBanner) return null;
+  if (!allBanner) return <Loading />;
   return (
     <div className={styles.container} id="banner">
       <div className={styles.header}>
@@ -26,68 +30,11 @@ const BannerManagement = () => {
         </Modal>
       </div>
 
-      <div className={styles.table_wrapper}>
-        <table className={styles.table}>
-          <thead>
-            <tr>
-              <th>ID</th>
-              <th>제목</th>
-              <th>링크</th>
-              <th>생성일</th>
-              <th className={styles.action_title}>Actions</th>
-            </tr>
-          </thead>
-
-          <tbody>
-            {allBanner?.items.map((banner) => {
-              const date = formatDate(
-                new Date(banner.createdAt),
-                "yyyy. MM. dd"
-              );
-
-              return (
-                <tr key={banner.bannerId}>
-                  <td>{banner.bannerId}</td>
-                  <td>{banner.title}</td>
-                  <td className={styles.banner_link}>
-                    <a href={banner.linkUrl} target="_blank">
-                      {banner.linkUrl}
-                    </a>
-                  </td>
-                  <td>{date}</td>
-                  <td>
-                    <div className={styles.actions}>
-                      <Modal>
-                        <ModalTrigger>
-                          <Button size="sm" variant="ghost">
-                            Edit
-                          </Button>
-                        </ModalTrigger>
-                        <ModalContent>
-                          <ModifyBanner item={banner} callback={updateBanner} />
-                        </ModalContent>
-                      </Modal>
-
-                      <Modal>
-                        <ModalTrigger>
-                          <Button size="sm" variant="ghost">
-                            Delete
-                          </Button>
-                        </ModalTrigger>
-                        <ModalContent>
-                          <DeleteModal
-                            callback={() => deleteBanner(banner.bannerId)}
-                          />
-                        </ModalContent>
-                      </Modal>
-                    </div>
-                  </td>
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-      </div>
+      <BannerTable
+        banners={allBanner}
+        updateBanner={updateBanner}
+        deleteBanner={deleteBanner}
+      />
     </div>
   );
 };
