+++ src/pages/Admin/Comment/CommentManagement.jsx
@@ -1,23 +1,15 @@
-import { useAxios, usePagination } from "@/hooks";
-import styles from "./CommentManagement.module.css";
-import { useCallback, useEffect } from "react";
+import { Loading, Pagination } from "@/components";
 import { APIEndPoints } from "@/constants";
-import {
-  Button,
-  DeleteModal,
-  Modal,
-  ModalContent,
-  ModalTrigger,
-  Pagination,
-} from "@/components";
-import { formatDate } from "date-fns";
-import { buildPath, handleApiCall } from "@/utils";
-import CommentModal from "./CommentModal";
 import { useToast } from "@/contexts";
+import { useAxios, usePagination } from "@/hooks";
+import { buildPath, handleApiCall } from "@/utils";
+import { useCallback, useEffect } from "react";
+import styles from "./CommentManagement.module.css";
+import CommentTable from "./CommentTable";
 
 const CommentManagement = () => {
   const { createToast } = useToast();
-  const { fetchData, response } = useAxios();
+  const { fetchData, response, loading } = useAxios();
   const { fetchData: deleteApi } = useAxios();
   const { currentPage, totalPage, setTotalPage, handlePageChange } =
     usePagination();
@@ -61,62 +53,16 @@ const CommentManagement = () => {
     });
   }, [currentPage, fetchData, setTotalPage]);
 
+  if (loading) return <Loading />;
   return (
     <div className={styles.container}>
       <p className={styles.title}>신고된 댓글 관리</p>
 
-      <div className={styles.table_wrapper}>
-        <table className={styles.table}>
-          <thead>
-            <tr>
-              <th>ID</th>
-              <th>내용</th>
-              <th>작성자ID</th>
-              <th>작성일</th>
-              <th>신고 수</th>
-              <th>
-                <p className={styles.action_title}>Actions</p>
-              </th>
-            </tr>
-          </thead>
-          <tbody>
-            {response?.items.map((comment) => {
-              const date = formatDate(
-                new Date(comment.createdAt),
-                "yyyy. MM. dd"
-              );
+      <CommentTable
+        comments={response?.items ?? []}
+        deleteComment={handleDelete}
+      />
 
-              return (
-                <tr key={comment.commentId}>
-                  <td>{comment.commentId}</td>
-                  <td>{comment.contents}</td>
-                  <td>{comment.user.userId}</td>
-                  <td>{date}</td>
-                  <td>{comment.reportCount}</td>
-                  <td>
-                    <div className={styles.actions}>
-                      <CommentModal comment={comment} />
-
-                      <Modal>
-                        <ModalTrigger>
-                          <Button size="sm" variant="ghost">
-                            Delete
-                          </Button>
-                        </ModalTrigger>
-                        <ModalContent>
-                          <DeleteModal
-                            callback={() => handleDelete(comment.commentId)}
-                          />
-                        </ModalContent>
-                      </Modal>
-                    </div>
-                  </td>
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-      </div>
       <div className={styles.pagination}>
         <Pagination
           currentPage={currentPage}
