+++ src/pages/Admin/Comment/CommentTable.module.css
@@ -1,24 +1,3 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-}
-
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 5rem;
-  box-sizing: border-box;
-  padding: 0 2rem;
-  box-shadow: var(--shadow-md);
-}
-
-.title {
-  font-size: var(--text-xl);
-  font-weight: var(--font-weight-semibold);
-}
-
 .table_wrapper {
   flex: 1;
   box-sizing: border-box;
@@ -54,9 +33,4 @@
 .actions {
   display: flex;
   gap: 0.5rem;
-}
-
-.pagination {
-  display: flex;
-  justify-content: center;
 }
