+++ src/pages/Admin/User/AllUserTable.module.css
@@ -1,0 +1,36 @@
+.table_wrapper {
+  flex: 1;
+  box-sizing: border-box;
+  padding: 1rem 2rem;
+}
+
+.table {
+  width: 100%;
+  border-collapse: collapse;
+}
+
+.table th,
+.table td {
+  padding: 0.8rem;
+  text-align: left;
+  border-bottom: 1px solid var(--color-gray-300);
+
+  max-width: 10rem;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.table th {
+  font-weight: var(--font-weight-semibold);
+}
+
+.action_title {
+  position: relative;
+  left: 0.5rem;
+}
+
+.actions {
+  display: flex;
+  gap: 0.5rem;
+}
